requirements.txt
requests
Django
djangorestframework